en train de gerer les listes

+++ libft/srcs/lst/printlst.c
@@ -4,7 +4,7 @@ void	printlst_int(t_list_int *lst)
 {
 	while (lst != NULL)
 	{
-		printf("%d\n", lst->content);
+		printf("%d\n", (long int)lst->content);
 		lst = lst->next;
 	}
 }
